+++ CHANGELOG.rst
@@ -18,6 +18,8 @@ of the list).
 DrizzlePac (DEVELOPMENT)
 ========================
 
+- Updated code that relies on ``tweakwcs`` to use new API. [#220]
+
 - Added hlapipeline alignment code to package. [#216]
 
 - Update runastrodriz to compute and apply absolute astrometric corrections to